move examples code to lib
