+++ foundationdb/src/cluster.rs
@@ -1,0 +1,71 @@
+use foundationdb_sys as fdb;
+use future::*;
+use futures::{Async, Future};
+use std;
+use std::sync::Arc;
+
+use database::*;
+use error::*;
+
+#[derive(Clone)]
+pub struct Cluster {
+    inner: Arc<ClusterInner>,
+}
+impl Cluster {
+    pub fn new(path: &str) -> ClusterGet {
+        let path_str = std::ffi::CString::new(path).unwrap();
+        let f = unsafe { fdb::fdb_create_cluster(path_str.as_ptr()) };
+        ClusterGet {
+            inner: FdbFuture::new(f),
+        }
+    }
+
+    //TODO: impl Future
+    pub fn create_database(&self) -> Box<Future<Item = Database, Error = FdbError>> {
+        let f = unsafe {
+            let f_db = fdb::fdb_cluster_create_database(self.inner.inner, b"DB" as *const _, 2);
+            let cluster = self.clone();
+            FdbFuture::new(f_db)
+                .and_then(|f| f.get_database())
+                .map(|db| Database::new(cluster, db))
+        };
+        Box::new(f)
+    }
+}
+
+pub struct ClusterGet {
+    inner: FdbFuture,
+}
+impl Future for ClusterGet {
+    type Item = Cluster;
+    type Error = FdbError;
+
+    fn poll(&mut self) -> Result<Async<Self::Item>> {
+        match self.inner.poll() {
+            Ok(Async::Ready(r)) => match unsafe { r.get_cluster() } {
+                Ok(c) => Ok(Async::Ready(Cluster {
+                    inner: Arc::new(ClusterInner::new(c)),
+                })),
+                Err(e) => Err(e),
+            },
+            Ok(Async::NotReady) => Ok(Async::NotReady),
+            Err(e) => Err(e),
+        }
+    }
+}
+
+struct ClusterInner {
+    inner: *mut fdb::FDBCluster,
+}
+impl ClusterInner {
+    fn new(inner: *mut fdb::FDBCluster) -> Self {
+        Self { inner }
+    }
+}
+impl Drop for ClusterInner {
+    fn drop(&mut self) {
+        unsafe {
+            fdb::fdb_cluster_destroy(self.inner);
+        }
+    }
+}
